refactored how the client requests and stores reviews so that the client can do less work. majority of the work happens on the server side now.
also - there's something strange going on with the API where there is a mismatch between reviews and ratings. That's not supposed to be possible per the spec, and it's messing up the conditional rendering to tell if 'more reviews' button should be available or not. I have a rather poor work around in place right now so the more reviews will disappear if the server doesn't send any review back. i'm reaching out to Alex for clarification on whether we need to accommodate having a mismatch between reviews/ratings (deviation from spec) or not

will come back and fix the workaround based on Alex's answer